requirements.txt
streamlit
pandas
numpy
sqlalchemy
mysql-connector-python
scikit-learn
xgboost
category-encoders
matplotlib
plotly
shap
joblib
python-dotenv